requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==25.1.0
blinker==1.7.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
diff-match-patch==20241021
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-import-export==4.3.7
django_csp==3.8
djangorestframework==3.15.2
et_xmlfile==2.0.0
exceptiongroup==1.1.1
firebase-admin==6.6.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.2
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.1.0
multidict==6.1.0
mypy_extensions==1.1.0
numpy==2.2.1
openpyxl==3.1.5
outcome==1.2.0
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.7
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pyfcm==2.0.7
PyJWT==2.10.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
robotframework==6.1.1
robotframework-pythonlibcore==4.2.0
robotframework-requests==0.9.5
robotframework-seleniumlibrary==6.1.2
rsa==4.9
selenium==4.11.2
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
sqlparse==0.5.1
tablib==3.8.0
telnetlib3==2.0.2
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.3
webdriver-manager==4.0.1
Werkzeug==3.0.1
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.18.3
